Moved @wasmer/wasmer-pack integration tests into the wasm dir
