+++ crates/wasm/Cargo.toml
@@ -20,6 +20,10 @@ bindings = { wai-version = "0.2.0", exports = "wasmer-pack.exports.wai" }
 [lib]
 crate-type = ["cdylib", "rlib"]
 
+[[test]]
+name = "wasmer-pack-wasm-integration-tests"
+harness = false
+
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
@@ -27,3 +31,8 @@ anyhow = "1.0.62"
 getrandom = { version = "0.2.8", features = ["custom"] }
 original = { package = "wasmer-pack", version = "0.6.0", path = "../wasmer-pack" }
 wai-bindgen-rust = "0.2.1"
+
+[dev-dependencies]
+anyhow = "1.0.66"
+tracing-subscriber = { version = "0.3.16", features = ["env-filter"] }
+wasmer-pack-testing = { path = "../testing" }
