+++ crates/wasm/jest.config.js
@@ -1,0 +1,6 @@
+/** @type {import('jest').Config} */
+const config = {
+  verbose: true,
+  testPathIgnorePatterns: ["node_modules", "snapshots"],
+};
+module.exports = config;
