--- crates/wasmer-pack/tests/js/.prettierrc.json
@@ -1,5 +1,0 @@
-{
-    "semi": true,
-    "singleQuote": false,
-    "tabWidth": 4
-}
