--- crates/wasmer-pack/tests/js/jasmine.json
@@ -1,11 +1,0 @@
-{
-    "reporters": [
-        {
-            "name": "jasmine-spec-reporter#SpecReporter",
-            "options": {
-                "displayStacktrace": "all"
-            }
-        }
-    ],
-    "spec_files": ["**/*.test.ts"]
-}
