--- crates/wasmer-pack/tests/js/package.json
@@ -1,22 +1,0 @@
-{
-    "name": "integration-tests",
-    "private": true,
-    "type": "module",
-    "scripts": {
-        "fmt": "prettier --write .",
-        "test": "cross-env NODE_OPTIONS='--loader=ts-node/esm --no-warnings' jasmine --config=jasmine.json"
-    },
-    "dependencies": {
-        "@wasmer/wasi": "^1.1.2",
-        "@wasmer/wasmer-pack": "file:../../../../target/tmp/wasmer-pack/javascript/package"
-    },
-    "devDependencies": {
-        "@types/jasmine": "^4.3.0",
-        "@types/node": "^18.7.6",
-        "cross-env": "^7.0.3",
-        "jasmine": "^4.4.0",
-        "prettier": "^2.7.1",
-        "ts-node": "^10.9.1",
-        "typescript": "^4.8.3"
-    }
-}
